# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Saffe — Security analysis API for vibe-coded web apps
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface DomainVerification {
  id: string;
  userId: string;
  domain: string;
  token: string;
  verified: boolean;
  /** @nullable */
  verifiedAt?: string | null;
  createdAt: string;
}

export interface InitVerificationRequest {
  url: string;
}

export interface InitVerificationResponse {
  domain: string;
  token: string;
  metaTag: string;
  alreadyVerified?: boolean;
}

export interface ConfirmVerificationRequest {
  domain: string;
}

export interface ConfirmVerificationResponse {
  verified: boolean;
  domain: string;
  message?: string;
}

export type ScanResultSeverity =
  (typeof ScanResultSeverity)[keyof typeof ScanResultSeverity];

export const ScanResultSeverity = {
  info: "info",
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface ScanResult {
  id: string;
  checkType: string;
  severity: ScanResultSeverity;
  title: string;
  description: string;
  /** @nullable */
  details?: string | null;
  /** @nullable */
  fixSuggestion?: string | null;
  /** @nullable */
  vibePrompt?: string | null;
  passed: boolean;
}

export type ScanStatus = (typeof ScanStatus)[keyof typeof ScanStatus];

export const ScanStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

/**
 * @nullable
 */
export type ScanRiskLevel =
  | (typeof ScanRiskLevel)[keyof typeof ScanRiskLevel]
  | null;

export const ScanRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Scan {
  id: string;
  userId: string;
  url: string;
  domain: string;
  status: ScanStatus;
  /** @nullable */
  riskLevel: ScanRiskLevel;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
  results?: ScanResult[];
}

export interface CreateScanRequest {
  url: string;
}

export interface CreateScanResponse {
  scanId: string;
  status: string;
}

export interface ListScansResponse {
  scans: Scan[];
}

export interface SendReportEmailRequest {
  scanId: string;
  email: string;
}

export interface SendReportEmailResponse {
  success: boolean;
  message?: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  /**
   * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
   */
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};