# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Saffe — Security analysis API for vibe-coded web apps
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
  - name: verification
    description: Domain ownership verification
  - name: scans
    description: Security scan operations

components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']

    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'

    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1

    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string

    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true

    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    DomainVerification:
      type: object
      required: [id, userId, domain, token, verified, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        domain:
          type: string
        token:
          type: string
        verified:
          type: boolean
        verifiedAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time

    InitVerificationRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri

    InitVerificationResponse:
      type: object
      required: [domain, token, metaTag]
      properties:
        domain:
          type: string
        token:
          type: string
        metaTag:
          type: string
        alreadyVerified:
          type: boolean

    ConfirmVerificationRequest:
      type: object
      required: [domain]
      properties:
        domain:
          type: string

    ConfirmVerificationResponse:
      type: object
      required: [verified, domain]
      properties:
        verified:
          type: boolean
        domain:
          type: string
        message:
          type: string

    ScanResult:
      type: object
      required: [id, checkType, severity, title, description, passed]
      properties:
        id:
          type: string
        checkType:
          type: string
        severity:
          type: string
          enum: [info, low, medium, high, critical]
        title:
          type: string
        description:
          type: string
        details:
          type: [string, 'null']
        fixSuggestion:
          type: [string, 'null']
        vibePrompt:
          type: [string, 'null']
        passed:
          type: boolean

    Scan:
      type: object
      required: [id, userId, url, domain, status, riskLevel, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        url:
          type: string
        domain:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        riskLevel:
          type: [string, 'null']
          enum: [low, medium, high, critical, null]
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: [string, 'null']
          format: date-time
        results:
          type: array
          items:
            $ref: '#/components/schemas/ScanResult'

    CreateScanRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri

    CreateScanResponse:
      type: object
      required: [scanId, status]
      properties:
        scanId:
          type: string
        status:
          type: string

    ListScansResponse:
      type: object
      required: [scans]
      properties:
        scans:
          type: array
          items:
            $ref: '#/components/schemas/Scan'

    SendReportEmailRequest:
      type: object
      required: [scanId, email]
      properties:
        scanId:
          type: string
        email:
          type: string
          format: email

    SendReportEmailResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'

  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Relative path to redirect to after login (must start with `/`). Defaults to `/`.
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.

  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the `returnTo` path on success (defaults to `/`) or `/api/login` on failure.

  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.

  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

  /verify/init:
    post:
      tags: [verification]
      operationId: initDomainVerification
      summary: Initialize domain ownership verification
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitVerificationRequest'
      responses:
        '200':
          description: Verification token generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitVerificationResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /verify/confirm:
    post:
      tags: [verification]
      operationId: confirmDomainVerification
      summary: Confirm domain ownership by checking meta tag
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmVerificationRequest'
      responses:
        '200':
          description: Verification status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmVerificationResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /scans:
    get:
      tags: [scans]
      operationId: listScans
      summary: List all scans for the current user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: List of scans
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListScansResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

    post:
      tags: [scans]
      operationId: createScan
      summary: Start a security scan for a URL
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScanRequest'
      responses:
        '201':
          description: Scan created and started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateScanResponse'
        '400':
          description: Bad request or domain not verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /scans/{scanId}:
    get:
      tags: [scans]
      operationId: getScan
      summary: Get a specific scan with results
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: scanId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Scan details with results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scan'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Scan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /scans/{scanId}/send-email:
    post:
      tags: [scans]
      operationId: sendScanReport
      summary: Send scan report by email
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: scanId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendReportEmailRequest'
      responses:
        '200':
          description: Email sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendReportEmailResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Scan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'