# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Saffe — Security analysis API for vibe-coded web apps
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get the currently authenticated user
 */
export const GetCurrentAuthUserHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

export const GetCurrentAuthUserResponse = zod.object({
  user: zod.union([
    zod.object({
      id: zod.string(),
      email: zod.string().email().nullable(),
      firstName: zod.string().nullable(),
      lastName: zod.string().nullable(),
      profileImageUrl: zod.string().nullable(),
    }),
    zod.null(),
  ]),
});

/**
 * @summary Start the browser OIDC login flow
 */
export const BeginBrowserLoginQueryParams = zod.object({
  returnTo: zod.coerce
    .string()
    .optional()
    .describe(
      "Relative path to redirect to after login (must start with `\/`). Defaults to `\/`.",
    ),
});

/**
 * @summary Complete the browser OIDC login flow
 */
export const HandleBrowserLoginCallbackQueryParams = zod.object({
  code: zod.coerce.string().optional(),
  state: zod.coerce.string().optional(),
  iss: zod.coerce.string().url().optional(),
});

/**
 * @summary Clear the session and begin OIDC logout
 */
export const LogoutBrowserSessionHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

/**
 * @summary Exchange a mobile OIDC code for a session token
 */

export const ExchangeMobileAuthorizationCodeBody = zod.object({
  code: zod.string().min(1),
  code_verifier: zod.string().min(1),
  redirect_uri: zod.string().url().min(1),
  state: zod.string().min(1),
  nonce: zod.string().min(1).optional(),
});

export const ExchangeMobileAuthorizationCodeResponse = zod.object({
  token: zod.string(),
});

/**
 * @summary Delete a mobile session token
 */
export const LogoutMobileSessionHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

export const LogoutMobileSessionResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Initialize domain ownership verification
 */
export const InitDomainVerificationHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

export const InitDomainVerificationBody = zod.object({
  url: zod.string().url(),
});

export const InitDomainVerificationResponse = zod.object({
  domain: zod.string(),
  token: zod.string(),
  metaTag: zod.string(),
  alreadyVerified: zod.boolean().optional(),
});

/**
 * @summary Confirm domain ownership by checking meta tag
 */
export const ConfirmDomainVerificationHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

export const ConfirmDomainVerificationBody = zod.object({
  domain: zod.string(),
});

export const ConfirmDomainVerificationResponse = zod.object({
  verified: zod.boolean(),
  domain: zod.string(),
  message: zod.string().optional(),
});

/**
 * @summary List all scans for the current user
 */
export const ListScansHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

export const ListScansResponse = zod.object({
  scans: zod.array(
    zod.object({
      id: zod.string(),
      userId: zod.string(),
      url: zod.string(),
      domain: zod.string(),
      status: zod.enum(["pending", "running", "completed", "failed"]),
      riskLevel: zod
        .union([
          zod.literal("low"),
          zod.literal("medium"),
          zod.literal("high"),
          zod.literal("critical"),
          zod.literal(null),
        ])
        .nullable(),
      createdAt: zod.date(),
      completedAt: zod.date().nullish(),
      results: zod
        .array(
          zod.object({
            id: zod.string(),
            checkType: zod.string(),
            severity: zod.enum(["info", "low", "medium", "high", "critical"]),
            title: zod.string(),
            description: zod.string(),
            details: zod.string().nullish(),
            fixSuggestion: zod.string().nullish(),
            vibePrompt: zod.string().nullish(),
            passed: zod.boolean(),
          }),
        )
        .optional(),
    }),
  ),
});

/**
 * @summary Start a security scan for a URL
 */
export const CreateScanHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

export const CreateScanBody = zod.object({
  url: zod.string().url(),
});

/**
 * @summary Get a specific scan with results
 */
export const GetScanParams = zod.object({
  scanId: zod.coerce.string(),
});

export const GetScanHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

export const GetScanResponse = zod.object({
  id: zod.string(),
  userId: zod.string(),
  url: zod.string(),
  domain: zod.string(),
  status: zod.enum(["pending", "running", "completed", "failed"]),
  riskLevel: zod
    .union([
      zod.literal("low"),
      zod.literal("medium"),
      zod.literal("high"),
      zod.literal("critical"),
      zod.literal(null),
    ])
    .nullable(),
  createdAt: zod.date(),
  completedAt: zod.date().nullish(),
  results: zod
    .array(
      zod.object({
        id: zod.string(),
        checkType: zod.string(),
        severity: zod.enum(["info", "low", "medium", "high", "critical"]),
        title: zod.string(),
        description: zod.string(),
        details: zod.string().nullish(),
        fixSuggestion: zod.string().nullish(),
        vibePrompt: zod.string().nullish(),
        passed: zod.boolean(),
      }),
    )
    .optional(),
});

/**
 * @summary Send scan report by email
 */
export const SendScanReportParams = zod.object({
  scanId: zod.coerce.string(),
});

export const SendScanReportHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

export const SendScanReportBody = zod.object({
  scanId: zod.string(),
  email: zod.string().email(),
});

export const SendScanReportResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});