# lib/api-zod/src/generated/types/confirmVerificationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Saffe — Security analysis API for vibe-coded web apps
 * OpenAPI spec version: 0.1.0
 */

export interface ConfirmVerificationRequest {
  domain: string;
}